// typechain/PProxyOverridable.d.ts
interface PProxyOverridableInterface extends Interface {
  functions: {
    addressToBytes32: TypedFunctionDescription<{
      encode([_value]: [string]): string;
    }>;

    bytes32ToAddress: TypedFunctionDescription<{
      encode([_value]: [Arrayish]): string;
    }>;

    getImplementation: TypedFunctionDescription<{ encode([]: []): string }>;

    getOverrides: TypedFunctionDescription<{ encode([]: []): string }>;

    getProxyOwner: TypedFunctionDescription<{ encode([]: []): string }>;

    readAddress: TypedFunctionDescription<{
      encode([_key]: [Arrayish]): string;
    }>;

    readBool: TypedFunctionDescription<{ encode([_key]: [Arrayish]): string }>;

    setImplementation: TypedFunctionDescription<{
      encode([_newImplementation]: [string]): string;
    }>;

    setOverrides: TypedFunctionDescription<{
      encode([_newOverrides]: [string]): string;
    }>;

    setProxyOwner: TypedFunctionDescription<{
      encode([_newOwner]: [string]): string;
    }>;

    storageRead: TypedFunctionDescription<{
      encode([_key]: [Arrayish]): string;
    }>;
  };

  events: {};
}

export class PProxyOverridable extends Contract {
  connect(signerOrProvider: Signer | Provider | string): PProxyOverridable;
  attach(addressOrName: string): PProxyOverridable;
  deployed(): Promise<PProxyOverridable>;

  on(event: EventFilter | string, listener: Listener): PProxyOverridable;
  once(event: EventFilter | string, listener: Listener): PProxyOverridable;
  addListener(
    eventName: EventFilter | string,
    listener: Listener
  ): PProxyOverridable;
  removeAllListeners(eventName: EventFilter | string): PProxyOverridable;
  removeListener(eventName: any, listener: Listener): PProxyOverridable;

  interface: PProxyOverridableInterface;

  functions: {
    addressToBytes32(_value: string): Promise<string>;

    bytes32ToAddress(_value: Arrayish): Promise<string>;

    getImplementation(): Promise<string>;

    getOverrides(): Promise<string>;

    getProxyOwner(): Promise<string>;

    readAddress(_key: Arrayish): Promise<string>;

    readBool(_key: Arrayish): Promise<boolean>;

    setImplementation(
      _newImplementation: string,
      overrides?: TransactionOverrides
    ): Promise<ContractTransaction>;

    setOverrides(
      _newOverrides: string,
      overrides?: TransactionOverrides
    ): Promise<ContractTransaction>;

    setProxyOwner(
      _newOwner: string,
      overrides?: TransactionOverrides
    ): Promise<ContractTransaction>;

    storageRead(_key: Arrayish): Promise<string>;
  };

  addressToBytes32(_value: string): Promise<string>;

  bytes32ToAddress(_value: Arrayish): Promise<string>;

  getImplementation(): Promise<string>;

  getOverrides(): Promise<string>;

  getProxyOwner(): Promise<string>;

  readAddress(_key: Arrayish): Promise<string>;

  readBool(_key: Arrayish): Promise<boolean>;

  setImplementation(
    _newImplementation: string,
    overrides?: TransactionOverrides
  ): Promise<ContractTransaction>;

  setOverrides(
    _newOverrides: string,
    overrides?: TransactionOverrides
  ): Promise<ContractTransaction>;

  setProxyOwner(
    _newOwner: string,
    overrides?: TransactionOverrides
  ): Promise<ContractTransaction>;

  storageRead(_key: Arrayish): Promise<string>;

  filters: {};

  estimate: {
    addressToBytes32(_value: string): Promise<BigNumber>;

    bytes32ToAddress(_value: Arrayish): Promise<BigNumber>;

    getImplementation(): Promise<BigNumber>;

    getOverrides(): Promise<BigNumber>;

    getProxyOwner(): Promise<BigNumber>;

    readAddress(_key: Arrayish): Promise<BigNumber>;

    readBool(_key: Arrayish): Promise<BigNumber>;

    setImplementation(_newImplementation: string): Promise<BigNumber>;

    setOverrides(_newOverrides: string): Promise<BigNumber>;

    setProxyOwner(_newOwner: string): Promise<BigNumber>;

    storageRead(_key: Arrayish): Promise<BigNumber>;
  };
}