// typechain/TestOverrides.d.ts
/* Generated by ts-generator ver. 0.0.8 */
/* tslint:disable */

import { Contract, ContractTransaction, EventFilter, Signer } from "ethers";
import { Listener, Provider } from "ethers/providers";
import { Arrayish, BigNumber, BigNumberish, Interface } from "ethers/utils";
import {
  TransactionOverrides,
  TypedEventDescription,
  TypedFunctionDescription
} from ".";

interface TestOverridesInterface extends Interface {
  functions: {
    doesOverride: TypedFunctionDescription<{
      encode([_selector]: [Arrayish]): string;
    }>;

    name: TypedFunctionDescription<{ encode([]: []): string }>;

    setValue1: TypedFunctionDescription<{ encode([_value]: [string]): string }>;

    symbol: TypedFunctionDescription<{ encode([]: []): string }>;

    value: TypedFunctionDescription<{ encode([]: []): string }>;

    value1: TypedFunctionDescription<{ encode([]: []): string }>;
  };

  events: {};
}

export class TestOverrides extends Contract {
  connect(signerOrProvider: Signer | Provider | string): TestOverrides;
  attach(addressOrName: string): TestOverrides;
  deployed(): Promise<TestOverrides>;

  on(event: EventFilter | string, listener: Listener): TestOverrides;
  once(event: EventFilter | string, listener: Listener): TestOverrides;
  addListener(
    eventName: EventFilter | string,
    listener: Listener
  ): TestOverrides;
  removeAllListeners(eventName: EventFilter | string): TestOverrides;
  removeListener(eventName: any, listener: Listener): TestOverrides;

  interface: TestOverridesInterface;

  functions: {
    doesOverride(_selector: Arrayish): Promise<boolean>;

    name(): Promise<string>;

    setValue1(
      _value: string,
      overrides?: TransactionOverrides
    ): Promise<ContractTransaction>;

    symbol(): Promise<string>;

    value(): Promise<string>;

    value1(): Promise<string>;
  };

  doesOverride(_selector: Arrayish): Promise<boolean>;

  name(): Promise<string>;

  setValue1(
    _value: string,
    overrides?: TransactionOverrides
  ): Promise<ContractTransaction>;

  symbol(): Promise<string>;

  value(): Promise<string>;

  value1(): Promise<string>;

  filters: {};

  estimate: {
    doesOverride(_selector: Arrayish): Promise<BigNumber>;

    name(): Promise<BigNumber>;

    setValue1(_value: string): Promise<BigNumber>;

    symbol(): Promise<BigNumber>;

    value(): Promise<BigNumber>;

    value1(): Promise<BigNumber>;
  };
}